+++ source/DSCResources/MSFT_AdfsFarm/MSFT_AdfsFarm.psm1
@@ -101,13 +101,16 @@ function Get-TargetResource
         [System.String]
         $FederationServiceName,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.String]
         $CertificateThumbprint,
 
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
-        $Credential
+        $Credential,
+
+        [System.String]
+        $CertificateName
     )
 
     # Set Verbose and Debug parameters
@@ -156,6 +159,7 @@ function Get-TargetResource
         if ($adfsSslCertificate)
         {
             $certificateThumbprint = $adfsSslCertificate.CertificateHash
+            $certificateName = $adfsSslCertificate.HostName
         }
         else
         {
@@ -207,6 +211,7 @@ function Get-TargetResource
         $returnValue = @{
             FederationServiceName         = $adfsProperties.HostName
             CertificateThumbprint         = $certificateThumbprint
+            CertificateName               = $certificateName
             FederationServiceDisplayName  = $adfsProperties.DisplayName
             GroupServiceAccountIdentifier = $groupServiceAccountIdentifier
             ServiceAccountCredential      = $serviceAccountCredential
@@ -222,6 +227,7 @@ function Get-TargetResource
         $returnValue = @{
             FederationServiceName         = $FederationServiceName
             CertificateThumbprint         = $CertificateThumbprint
+            CertificateName               = $certificateName
             FederationServiceDisplayName  = $null
             GroupServiceAccountIdentifier = $null
             ServiceAccountCredential      = $null
@@ -274,7 +280,7 @@ function Set-TargetResource
         [System.String]
         $FederationServiceName,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.String]
         $CertificateThumbprint,
 
@@ -304,7 +310,16 @@ function Set-TargetResource
 
         [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance[]]
-        $AdminConfiguration
+        $AdminConfiguration,
+
+        [System.String]
+        $CertificateName,
+
+        [System.String]
+        $SigningCertificateName,
+
+        [System.String]
+        $DecryptionCertificateName
     )
 
     Write-Verbose -Message ($script:localizedData.SettingResourceMessage -f $FederationServiceName)
@@ -325,14 +340,31 @@ function Set-TargetResource
     if (-not $PSBoundParameters.ContainsKey('ServiceAccountCredential') -and
         -not $PSBoundParameters.ContainsKey('GroupServiceAccountIdentifier'))
     {
-        $errorMessage = $script:localizedData.ResourceMissingCredentialErrorMessage -f $FederationServiceName
+        $errorMessage = $script:localizedData.ResourceMissingCredentialErrorMessage
         New-InvalidArgumentException -Message $errorMessage -ArgumentName 'ServiceAccountCredential'
+    }
+
+    # Check whether both service certificate parameters have been specified
+    if ($PSBoundParameters.ContainsKey('CertificateThumbprint') -and
+        $PSBoundParameters.ContainsKey('CertificateName'))
+    {
+        $errorMessage = $script:localizedData.ResourceDuplicateServiceCertificateErrorMessage
+        New-InvalidArgumentException -Message $errorMessage -ArgumentName 'CertificateThumbprint'
+    }
+
+    # Check whether no service certificate parameters have been specified
+    if (-not $PSBoundParameters.ContainsKey('CertificateThumbprint') -and
+        -not $PSBoundParameters.ContainsKey('CertificateName'))
+    {
+        $errorMessage = $script:localizedData.ResourceMissingCredentialErrorMessage
+        New-InvalidArgumentException -Message $errorMessage -ArgumentName 'CertificateThumbprint'
     }
 
     $GetTargetResourceParms = @{
         FederationServiceName = $FederationServiceName
         CertificateThumbprint = $CertificateThumbprint
         Credential            = $Credential
+        CertificateName       = $CertificateName
     }
     $targetResource = Get-TargetResource @GetTargetResourceParms
 
@@ -356,6 +388,34 @@ function Set-TargetResource
             }
 
             $parameters.AdminConfiguration = $adminConfigurationHashTable
+        }
+
+        #YVAND
+        if ($PSBoundParameters.ContainsKey('CertificateName'))
+        {
+            Write-Verbose -Message "YVAND getting certificate $CertificateName"
+            $siteCert = Get-ChildItem -Path "cert:\LocalMachine\My\" -DnsName $CertificateName
+            $parameters.CertificateThumbprint = $siteCert.Thumbprint
+            $parameters.Remove('CertificateName')
+            Write-Verbose -Message "YVAND got certificate $CertificateName with Thumbprint $($siteCert.Thumbprint)"
+        }
+
+        if ($PSBoundParameters.ContainsKey('SigningCertificateName'))
+        {
+            Write-Verbose -Message "YVAND getting certificate $SigningCertificateName"
+            $signingCert = Get-ChildItem -Path "cert:\LocalMachine\My\" -DnsName $SigningCertificateName
+            $parameters.Add("SigningCertificateThumbprint", $signingCert.Thumbprint)
+            $parameters.Remove('SigningCertificateName')
+            Write-Verbose -Message "YVAND got certificate $SigningCertificateName with Thumbprint $($signingCert.Thumbprint)"
+        }
+
+        if ($PSBoundParameters.ContainsKey('DecryptionCertificateName'))
+        {
+            Write-Verbose -Message "YVAND getting certificate $DecryptionCertificateName"
+            $decryptionCert = Get-ChildItem -Path "cert:\LocalMachine\My\" -DnsName $DecryptionCertificateName
+            $parameters.Add("DecryptionCertificateThumbprint", $decryptionCert.Thumbprint)
+            $parameters.Remove('DecryptionCertificateName')
+            Write-Verbose -Message "YVAND got certificate $DecryptionCertificateName with Thumbprint $($decryptionCert.Thumbprint)"
         }
 
         try
@@ -416,7 +476,7 @@ function Test-TargetResource
         [System.String]
         $FederationServiceName,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.String]
         $CertificateThumbprint,
 
@@ -446,7 +506,16 @@ function Test-TargetResource
 
         [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance[]]
-        $AdminConfiguration
+        $AdminConfiguration,
+
+        [System.String]
+        $CertificateName,
+
+        [System.String]
+        $SigningCertificateName,
+
+        [System.String]
+        $DecryptionCertificateName
     )
 
     Write-Verbose -Message ($script:localizedData.TestingResourceMessage -f $FederationServiceName)
@@ -455,6 +524,7 @@ function Test-TargetResource
         FederationServiceName = $FederationServiceName
         CertificateThumbprint = $CertificateThumbprint
         Credential            = $Credential
+        CertificateName       = $CertificateName
     }
     $targetResource = Get-TargetResource @getTargetResourceParms