+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- AdfsFarm
+  - Added parameters `CertificateName`, `SigningCertificateName` and `DecryptionCertificateName`.
+
 ## [1.1.0] - 2020-12-22
 
 ### Added